requirements.txt
boto==2.38.0
bz2file==0.98
click==4.0
decorator==3.4.2
gensim==0.11.1.post1
line-profiler==1.0
logging==0.4.9.6
lxml==3.4.4
matplotlib==1.4.3
mock==1.0.1
networkx==1.9.1
nltk==3.0.3
nose==1.3.7
numpy==1.9.2
pandas==0.16.2
Pillow==2.9.0
py==1.4.30
pyparsing==2.0.3
python-dateutil==2.4.2
python-docx==0.8.5
pytz==2015.4
scikit-image==0.11.3
scikit-learn==0.16.1
scipy==0.15.1
simplejson==3.7.3
six==1.9.0
smart-open==1.2.1
wheel==0.24.0